requirements_airflow.txt
pyspark==3.0.1
tweepy==3.9.0
kafka-python==2.0.2
findspark==1.4.2